requirements.txt
appnope==0.1.0
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.0.0
certifi==2017.4.17
cffi==1.10.0
configparser==3.5.0
constantly==15.1.0
cryptography==1.9
cssselect==1.0.1
decorator==4.0.11
entrypoints==0.2.2
enum34==1.1.6
functools32==3.2.3.post2
html5lib==0.999999999
idna==2.5
incremental==17.5.0
ipaddress==1.0.18
ipykernel==4.6.1
ipython==5.4.1
ipython-genutils==0.2.0
isodate==0.5.4
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.0.1
jupyter-core==4.3.0
lxml==3.7.3
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.3.0
notebook==5.0.0
olefile==0.44
pandocfilters==1.4.1
parsel==1.2.0
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycparser==2.17
PyDispatcher==2.0.5
Pygments==2.2.0
pymongo==3.4.0
pyOpenSSL==17.0.0
python-dateutil==2.6.0
pyzmq==16.0.2
queuelib==1.4.2
scandir==1.5
Scrapy==1.4.0
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
slugify==0.0.1
terminado==0.6
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
Twisted==17.1.0
w3lib==1.17.0
wcwidth==0.1.7
webencodings==0.5.1
zope.interface==4.4.1
